pyproject.toml
[tool.poetry]
name = "dateflix-api"
version = "0.1.0"
description = ""
authors = ["Pythrick <pathrick@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = ""
beautifulsoup4 = "*"
dj-database-url = "*"
django-cors-headers = "*"
django-extensions = "*"
django-filter = "*"
django-heroku = "*"
djangorestframework = "*"
djangorestframework-simplejwt = "*"
dynaconf = "*"
gunicorn = "*"
httpx = "*"
psycopg2 = "*"
[tool.poetry.dev-dependencies]
pytest = "*"
pytest-django = "*"
pytest-cov = "*"
mypy = "*"
isort = "*"
flake8 = "*"
pre-commit = "^2.11.1"
[build-system]
requires = ["poetry>=1.1.5"]
build-backend = "poetry.masonry.api"
